decouple function node finder and collector from ParsedNodesByType
