+++ ecs.yaml
@@ -75,6 +75,8 @@ parameters:
 
         # @todo resolve!!!
         Symplify\CodingStandard\Sniffs\CleanCode\CognitiveComplexitySniff:
+            - 'src/NodeContainer/NodeCollector/ParsedFunctionLikeNodesByType.php'
+            - 'src/NodeContainer/NodeCollector/ParsedFunctionLikeNodeCollector.php'
             - 'packages/solid/src/Rector/ClassConst/PrivatizeLocalClassConstantRector.php'
             - 'packages/type-declaration/src/TypeInferer/PropertyTypeInferer/ConstructorPropertyTypeInferer.php'
             - 'packages/node-type-resolver/src/PhpDoc/NodeAnalyzer/DocBlockManipulator.php'
